# crypto/bn256: avoid silent memory allocations