… core/vm/contracts_test.go