```
$ benchcmp old.txt new.txt 

benchmark                                       old ns/op     new ns/op     delta
Bn256Add/chfast1-Gas=500-8                      51355         47504         -7.50%
Bn256Add/chfast2-Gas=500-8                      52026         50798         -2.36%
Bn256ScalarMul/chfast1-Gas=2000-8               520589        459350        -11.76%
Bn256ScalarMul/chfast2-Gas=2000-8               2010472       1805791       -10.18%
Bn256ScalarMul/chfast3-Gas=2000-8               1953982       1815215       -7.10%
Bn256ScalarMul/cdetrio1-Gas=2000-8              3068363       2848605       -7.16%
Bn256ScalarMul/cdetrio6-Gas=2000-8              3203281       2818420       -12.01%
Bn256ScalarMul/cdetrio11-Gas=2000-8             3122656       2954437       -5.39%
Bn256Pairing/jeff1-Gas=260000-8                 26528643      20770507      -21.71%
Bn256Pairing/jeff2-Gas=260000-8                 32784856      20312937      -38.04%
Bn256Pairing/jeff3-Gas=260000-8                 22369029      20904204      -6.55%
Bn256Pairing/jeff4-Gas=340000-8                 26514337      26958266      +1.67%
Bn256Pairing/jeff5-Gas=340000-8                 27133553      26003919      -4.16%
Bn256Pairing/jeff6-Gas=260000-8                 21695148      20670094      -4.72%
Bn256Pairing/empty_data-Gas=100000-8            1895591       1807918       -4.63%
Bn256Pairing/one_point-Gas=180000-8             15142486      14411274      -4.83%
Bn256Pairing/two_point_match_2-Gas=260000-8     13086229      12450534      -4.86%
Bn256Pairing/two_point_match_3-Gas=260000-8     21498610      20682219      -3.80%
Bn256Pairing/two_point_match_4-Gas=260000-8     21455913      20624358      -3.88%
Bn256Pairing/ten_point_match_1-Gas=900000-8     57057923      55680270      -2.41%
Bn256Pairing/ten_point_match_2-Gas=900000-8     65459519      65592808      +0.20%
Bn256Pairing/ten_point_match_3-Gas=260000-8     20439629      20492711      +0.26%

benchmark                                       old allocs     new allocs     delta
Bn256Add/chfast1-Gas=500-8                      113            110            -2.65%
Bn256Add/chfast2-Gas=500-8                      113            109            -3.54%
Bn256ScalarMul/chfast1-Gas=2000-8               1677           895            -46.63%
Bn256ScalarMul/chfast2-Gas=2000-8               6618           3367           -49.12%
Bn256ScalarMul/chfast3-Gas=2000-8               6592           3354           -49.12%
Bn256ScalarMul/cdetrio1-Gas=2000-8              10648          5379           -49.48%
Bn256ScalarMul/cdetrio6-Gas=2000-8              10623          5367           -49.48%
Bn256ScalarMul/cdetrio11-Gas=2000-8             10634          5373           -49.47%
Bn256Pairing/jeff1-Gas=260000-8                 57653          47287          -17.98%
Bn256Pairing/jeff2-Gas=260000-8                 57701          47305          -18.02%
Bn256Pairing/jeff3-Gas=260000-8                 57654          47276          -18.00%
Bn256Pairing/jeff4-Gas=340000-8                 73675          60411          -18.00%
Bn256Pairing/jeff5-Gas=340000-8                 73656          60397          -18.00%
Bn256Pairing/jeff6-Gas=260000-8                 57656          47283          -17.99%
Bn256Pairing/empty_data-Gas=100000-8            11446          11446          +0.00%
Bn256Pairing/one_point-Gas=180000-8             41240          33605          -18.51%
Bn256Pairing/two_point_match_2-Gas=260000-8     42683          36649          -14.14%
Bn256Pairing/two_point_match_3-Gas=260000-8     57205          46708          -18.35%
Bn256Pairing/two_point_match_4-Gas=260000-8     57223          46714          -18.36%
Bn256Pairing/ten_point_match_1-Gas=900000-8     166875         136762         -18.05%
Bn256Pairing/ten_point_match_2-Gas=900000-8     183124         149162         -18.55%
Bn256Pairing/ten_point_match_3-Gas=260000-8     57223          46714          -18.36%

benchmark                                       old bytes     new bytes     delta
Bn256Add/chfast1-Gas=500-8                      7338          6762          -7.85%
Bn256Add/chfast2-Gas=500-8                      7354          6666          -9.36%
Bn256ScalarMul/chfast1-Gas=2000-8               154672        68439         -55.75%
Bn256ScalarMul/chfast2-Gas=2000-8               620408        261993        -57.77%
Bn256ScalarMul/chfast3-Gas=2000-8               618297        261144        -57.76%
Bn256ScalarMul/cdetrio1-Gas=2000-8              991464        414940        -58.15%
Bn256ScalarMul/cdetrio6-Gas=2000-8              989814        414430        -58.13%
Bn256ScalarMul/cdetrio11-Gas=2000-8             990572        414715        -58.13%
Bn256Pairing/jeff1-Gas=260000-8                 3655095       2450243       -32.96%
Bn256Pairing/jeff2-Gas=260000-8                 3661326       2452354       -33.02%
Bn256Pairing/jeff3-Gas=260000-8                 3655692       2449049       -33.01%
Bn256Pairing/jeff4-Gas=340000-8                 4727586       3176081       -32.82%
Bn256Pairing/jeff5-Gas=340000-8                 4724646       3174094       -32.82%
Bn256Pairing/jeff6-Gas=260000-8                 3655195       2449380       -32.99%
Bn256Pairing/empty_data-Gas=100000-8            215304        215304        +0.00%
Bn256Pairing/one_point-Gas=180000-8             2461849       1605197       -34.80%
Bn256Pairing/two_point_match_2-Gas=260000-8     2106703       1431507       -32.05%
Bn256Pairing/two_point_match_3-Gas=260000-8     3526401       2327965       -33.98%
Bn256Pairing/two_point_match_4-Gas=260000-8     3529061       2329597       -33.99%
Bn256Pairing/ten_point_match_1-Gas=900000-8     9604107       6232946       -35.10%
Bn256Pairing/ten_point_match_2-Gas=900000-8     11531318      7621742       -33.90%
Bn256Pairing/ten_point_match_3-Gas=260000-8     3529058       2329614       -33.99%
```