@fjl 2017-09-25
This is being reviewed at https://go-review.googlesource.com/c/go/+/60671/